✨feat: 헤더 레이아웃 추가 및 적용
